+++ packages/module-workflow/src/client/NodeContext.ts
@@ -1,0 +1,9 @@
+import React, { useContext } from 'react';
+
+const ContextNode = React.createContext<any>({});
+
+export const ProviderContext = ContextNode.Provider;
+
+export function useContextNode() {
+  return useContext(ContextNode);
+}
